fixed bugs and expanded functionality

+++ javascope/DataProvider.java
@@ -1,8 +1,8 @@
 interface DataProvider {
     public void Update(String exp, int s);
     public String GetString(String in);
-    public double GetDouble(String in);
-    public double[] GetDoubleArray(String in);
+    public float GetFloat(String in);
+    public float[] GetFloatArray(String in);
     public String ErrorString();
     public boolean SupportsAsynch();
     }
